Split and reuse error parsing
