+++ liquidity/components/ContractError/package.json
@@ -5,7 +5,7 @@
   "version": "0.0.1",
   "dependencies": {
     "@chakra-ui/react": "^2.8.2",
-    "@snx-v3/useContractErrorParser": "workspace:*",
+    "@snx-v3/parseContractError": "workspace:*",
     "date-fns": "^2.30.0",
     "react": "^18.2.0"
   }
