--- liquidity/lib/parser/index.ts
@@ -1,1 +1,0 @@
-export * from './parseTxError';
