+++ liquidity/lib/withERC7412/withERC7412.ts
@@ -2,67 +2,23 @@
 import { EvmPriceServiceConnection } from '@pythnetwork/pyth-evm-js';
 import { offchainMainnetEndpoint, offchainTestnetEndpoint } from '@snx-v3/constants';
 import {
+  importAccountProxy,
   importAllErrors,
+  importClosePosition,
   importCoreProxy,
   importMulticall3,
   importPythERC7412Wrapper,
   importPythVerfier,
-  importClosePosition,
-  importAccountProxy,
   importUSDProxy,
 } from '@snx-v3/contracts';
+import { extractErrorData, PYTH_ERRORS } from '@snx-v3/parseContractError';
 import { notNil } from '@snx-v3/tsHelpers';
-import { Network, deploymentHasERC7412, getMagicProvider } from '@snx-v3/useBlockchain';
+import { deploymentHasERC7412, getMagicProvider, Network } from '@snx-v3/useBlockchain';
 import { ethers } from 'ethers';
 
 const IS_DEBUG =
   window.localStorage.getItem('DEBUG') === 'true' ||
   window.localStorage.DEBUG?.slice(0, 3) === 'snx';
-
-export const PYTH_ERRORS = [
-  // Function arguments are invalid (e.g., the arguments lengths mismatch)
-  // Signature: 0xa9cb9e0d
-  'error InvalidArgument()',
-  // Update data is coming from an invalid data source.
-  // Signature: 0xe60dce71
-  'error InvalidUpdateDataSource()',
-  // Update data is invalid (e.g., deserialization error)
-  // Signature: 0xe69ffece
-  'error InvalidUpdateData()',
-  // Insufficient fee is paid to the method.
-  // Signature: 0x025dbdd4
-  'error InsufficientFee()',
-  // There is no fresh update, whereas expected fresh updates.
-  // Signature: 0xde2c57fa
-  'error NoFreshUpdate()',
-  // There is no price feed found within the given range or it does not exists.
-  // Signature: 0x45805f5d
-  'error PriceFeedNotFoundWithinRange()',
-  // Price feed not found or it is not pushed on-chain yet.
-  // Signature: 0x14aebe68
-  'error PriceFeedNotFound()',
-  // Requested price is stale.
-  // Signature: 0x19abf40e
-  'error StalePrice()',
-  // Given message is not a valid Wormhole VAA.
-  // Signature: 0x2acbe915
-  'error InvalidWormholeVaa()',
-  // Governance message is invalid (e.g., deserialization error).
-  // Signature: 0x97363b35
-  'error InvalidGovernanceMessage()',
-  // Governance message is not for this contract.
-  // Signature: 0x63daeb77
-  'error InvalidGovernanceTarget()',
-  // Governance message is coming from an invalid data source.
-  // Signature: 0x360f2d87
-  'error InvalidGovernanceDataSource()',
-  // Governance message is old.
-  // Signature: 0x88d1b847
-  'error OldGovernanceMessage()',
-  // The wormhole address to set in SetWormholeAddress governance is invalid.
-  // Signature: 0x13d3ed82
-  'error InvalidWormholeAddressToSet()',
-];
 
 async function fetchOffchainData({
   priceIds,
@@ -182,17 +138,6 @@ async function logMulticall({
         return { $: 'unknown', data, value, ...rest };
       }
     })
-  );
-}
-
-export function extractErrorData(error: Error | any) {
-  return (
-    error?.error?.error?.error?.data ||
-    error?.error?.error?.data ||
-    error?.error?.data?.data ||
-    error?.error?.data ||
-    error?.data?.data ||
-    error?.data
   );
 }
